refactor: merge regional plot programs into single program
